# api/openapi.yaml
openapi: 3.0.1
info:
  title: Robot Delivery API
  version: 0.0.1
servers:
- url: http://localhost:8080/
...
paths:
...
  /robotAvailable:
    get:
      tags:
      - Robot
      description: Returns Robot availability
      operationId: getRobotAvailibility
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RobotAvailableResponse'
        "404":
          description: Robot not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Robot
    x-swagger-router-controller: robot
components:
  schemas:
...
    RobotAvailableResponse:
      type: object
      properties:
        robotId:
          type: number
        available:
          type: boolean
      example:
        available: true
        robotId: 123
...
    Error:
      required:
      - code
      type: object
      properties:
        code:
          type: string
...